// web/src/api/model/buildingResponseStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * nimap API
 * OpenAPI spec version: dev
 */

export type BuildingResponseStatus = typeof BuildingResponseStatus[keyof typeof BuildingResponseStatus];


export const BuildingResponseStatus = {
  planned: 'planned',
  active: 'active',
  decommissioned: 'decommissioned',
  archived: 'archived',
} as const;